feat(equipment service): add BL to update average rating of equipment, use decimal instead of double type for rating

+++ src/RentIT.ReviewsService/ReviewService.Core/DTO/Review/ReviewAddRequest.cs
@@ -5,6 +5,6 @@ namespace ReviewService.Core.DTO.Review;
 
 public record ReviewAddRequest(
     [Required] [MaxLength(1024)] string Description,
-    [Required] [Range(1, 5)] double Rating,
+    [Required] [Range(1, 5)] decimal Rating,
     [Required] Guid RentalId
     );
