+++ src/RentIT.ReviewsService/ReviewService.Core/DTO/Review/UserReviewResponse.cs
@@ -1,3 +1,3 @@
 ﻿namespace ReviewService.Core.DTO.Review;
 
-public record UserReviewResponse(Guid Id, string Description, double Rating);
+public record UserReviewResponse(Guid Id, string Description, decimal Rating);
