+++ src/RentIT.ReviewsService/ReviewService.Core/Domain/RepositoryContracts/IReviewRepository.cs
@@ -7,5 +7,5 @@ public interface IReviewRepository : IBaseReviewRepository
 {
     Task DeleteReviewAsync(Review review, CancellationToken cancellationToken = default);
 
-    Task<double?> GetReviewScoreAsync(Guid reviewId, CancellationToken cancellationToken = default);
+    Task<decimal?> GetReviewScoreAsync(Guid reviewId, CancellationToken cancellationToken = default);
 }
